Add Windows coloring
